added frontend

+++ scripts/uploading_S3.py
@@ -1,8 +1,22 @@
 import os
 import boto3
+import botocore
 
 #start S3 connection
 s3 = boto3.resource('s3')
+
+#check if file exists in S3
+def check_if_file_exists(sf):
+    try:
+        s3.Object('pablo-mushroom-bucket', sf).load()
+    except botocore.exceptions.ClientError as e:
+        if e.response['Error']['Code'] == "404":
+            return False
+        else:
+            raise
+    else:
+        return True
+
 #total number of files that we want to transfer
 total_files = sum([len(f) for r,d,f in os.walk('flat_images/')])
 print('uploading %i files to S3.' %total_files)
@@ -12,7 +26,7 @@ for (root, dirs, files) in os.walk('flat_images/'):
         sub_files = [root + '/' + w for w in files]
         print('Working on species %s - includes %i images' %(root, len(sub_files)))
         for sf in sub_files:
-            if os.path.isfile(sf):
+            if os.path.isfile(sf) and not check_if_file_exists(sf):
                 data = open(sf, 'rb')
                 s3.Bucket('pablo-mushroom-bucket').put_object(Key=sf, Body=data)
                 data.close()
